create sdl-splunk and sdl-jms module
