+++ .idea/encodings.xml
@@ -5,6 +5,8 @@
     <file url="file://$PROJECT_DIR$/sdl-core" charset="UTF-8" />
     <file url="file://$PROJECT_DIR$/sdl-core/src/main/resources" charset="UTF-8" />
     <file url="file://$PROJECT_DIR$/sdl-core/src/main/scala" charset="UTF-8" />
+    <file url="file://$PROJECT_DIR$/sdl-jms/src/main/scala" charset="UTF-8" />
     <file url="file://$PROJECT_DIR$/sdl-kafka/src/main/scala" charset="UTF-8" />
+    <file url="file://$PROJECT_DIR$/sdl-splunk/src/main/scala" charset="UTF-8" />
   </component>
 </project>
