+++ pom.xml
@@ -27,6 +27,8 @@
 	<modules>
 		<module>sdl-core</module>
 		<module>sdl-kafka</module>
+		<module>sdl-splunk</module>
+		<module>sdl-jms</module>
 	</modules>
 
 	<licenses>
@@ -496,6 +498,16 @@
 				<groupId>org.codehaus.mojo</groupId>
 				<artifactId>versions-maven-plugin</artifactId>
 				<version>2.7</version>
+			</plugin>
+
+			<!-- disable executing surefire tests as we use scalatest -->
+			<plugin>
+				<groupId>org.apache.maven.plugins</groupId>
+				<artifactId>maven-surefire-plugin</artifactId>
+				<version>2.12.4</version>
+				<configuration>
+					<skip>true</skip>
+				</configuration>
 			</plugin>
 
 		</plugins>
