+++ sdl-core/pom.xml
@@ -207,27 +207,9 @@
 		</dependency>
 
 		<dependency>
-			<groupId>com.jsuereth</groupId>
-			<artifactId>scala-arm_${scala.minor.version}</artifactId>
-			<version>${scala-arm.version}</version>
-		</dependency>
-
-		<dependency>
-			<groupId>com.splunk</groupId>
-			<artifactId>splunk</artifactId>
-			<version>${splunk.version}</version>
-		</dependency>
-
-		<dependency>
 			<groupId>org.scalaj</groupId>
 			<artifactId>scalaj-http_${scala.minor.version}</artifactId>
 			<version>2.3.0</version>
-		</dependency>
-
-		<dependency>
-			<groupId>javax.jms</groupId>
-			<artifactId>jms</artifactId>
-			<version>1.1</version>
 		</dependency>
 
 		<!-- Start Keycloak dependencies -->
