+++ sdl-jms/src/main/scala/io/smartdatalake/util/jms/TextMessageString.scala
