+++ sdl-jms/src/main/scala/io/smartdatalake/workflow/dataobject/JmsDataObject.scala
