+++ sdl-splunk/src/test/scala/io/smartdatalake/workflow/dataobject/SplunkDataObjectTest.scala
@@ -18,20 +18,75 @@
  */
 package io.smartdatalake.workflow.dataobject
 
+import java.time.format.DateTimeFormatter
 import java.time.{Duration, LocalDateTime}
 
 import com.splunk.{JobExportArgs, Service}
 import com.typesafe.config.ConfigFactory
+import io.smartdatalake.config.{ConfigParser, InstanceRegistry}
 import io.smartdatalake.config.SdlConfigObject.ConnectionId
 import io.smartdatalake.definitions.BasicAuthMode
 import io.smartdatalake.testutils.DataObjectTestSuite
 import io.smartdatalake.workflow.connection.{SplunkConnection, SplunkConnectionService}
 import org.apache.spark.sql.Row
-import org.scalatest.Assertions
+import org.scalatest.{Assertions, FlatSpec, Matchers}
 
 class SplunkDataObjectTest extends DataObjectTestSuite {
 
   val queryRawColumnNames: String = "\"_raw\", \"_time\""
+
+  test("SplunkDataObject is parsable") {
+    val fmt = DateTimeFormatter.ofPattern("yyyy-MM-dd HH:mm")
+    val from = LocalDateTime.parse("1970-01-01 00:00", fmt)
+    val to = LocalDateTime.parse("1970-01-01 00:01", fmt)
+    val duration = Duration.ofMinutes(1)
+
+    val config = ConfigFactory.parseString(
+      s"""
+         |connections = {
+         | con123 = {
+         |   type = SplunkConnection
+         |   host = test.host
+         |   port = 8080
+         |   auth-mode = {
+         |        type = BasicAuthMode
+         |        user-variable = "CLEAR#testuser"
+         |        password-variable = "CLEAR#secret"
+         |    }
+         | }
+         |}
+         |
+         |dataObjects = {
+         | 123 = {
+         |   type = SplunkDataObject
+         |   connectionId = con123
+         |   params = {
+         |     query = "round(3.5)"
+         |     queryFrom = "${from.format(fmt)}"
+         |     queryTo = "${to.format(fmt)}"
+         |     queryTimeInterval = 1
+         |     columnNames = [val1, val2]
+         |     parallelRequests = 100
+         |   }
+         | }
+         |}
+         |""".stripMargin).resolve
+    implicit val registry: InstanceRegistry = ConfigParser.parse(config)
+    val registry2 = new InstanceRegistry()
+    registry2.register(SplunkConnection("con123","test.host", 8080, BasicAuthMode("CLEAR#testuser", "CLEAR#secret")))
+    registry.getDataObjects.head shouldBe SplunkDataObject(
+      id = "123",
+      connectionId = "con123",
+      params = SplunkParams(
+        query = "round(3.5)",
+        queryFrom = from,
+        queryTo = to,
+        queryTimeInterval = duration,
+        columnNames = Seq("val1", "val2"),
+        parallelRequests = 100
+      )
+    )(registry2)
+  }
 
   test("splunk queries without table mappings should yield a dataframe having only _raw and _time fields") {
     // prepare
